aws_genai/pyproject.toml
[tool.poetry]
name = "adidas-aws-genai"
version = "0.1.0"
description = ""
authors = ["Kuthalingam <kaliappan.kuthalingam@capgemini.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "1.34.116"
streamlit = "1.35.0"
python-dotenv = "1.0.1"
langchain-aws = "0.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
